Move utils tests to sibling files
