+++ lib/colors.test.js
@@ -1,8 +1,8 @@
 'use strict'
 
 const { test } = require('tap')
-const getColorizerPrivate = require('../../lib/colors')
-const { colorizerFactory: getColorizerPublic } = require('../../index')
+const getColorizerPrivate = require('./colors')
+const { colorizerFactory: getColorizerPublic } = require('../index')
 
 const testDefaultColorizer = getColorizer => async t => {
   const colorizer = getColorizer()
