+++ lib/utils/index.js
@@ -3,7 +3,7 @@
 const { readdirSync } = require('fs')
 const { basename } = require('path')
 
-const files = readdirSync(__dirname)
+const files = readdirSync(__dirname).filter(f => f.endsWith('.test.js') === false)
 for (const file of files) {
   if (file === 'index.js') continue
