+++ lib/utils/prettify-metadata.test.js
@@ -1,0 +1,84 @@
+'use strict'
+
+const tap = require('tap')
+const prettifyMetadata = require('./prettify-metadata')
+
+tap.test('returns `undefined` if no metadata present', async t => {
+  const str = prettifyMetadata({ log: {} })
+  t.equal(str, undefined)
+})
+
+tap.test('works with only `name` present', async t => {
+  const str = prettifyMetadata({ log: { name: 'foo' } })
+  t.equal(str, '(foo)')
+})
+
+tap.test('works with only `pid` present', async t => {
+  const str = prettifyMetadata({ log: { pid: '1234' } })
+  t.equal(str, '(1234)')
+})
+
+tap.test('works with only `hostname` present', async t => {
+  const str = prettifyMetadata({ log: { hostname: 'bar' } })
+  t.equal(str, '(on bar)')
+})
+
+tap.test('works with only `name` & `pid` present', async t => {
+  const str = prettifyMetadata({ log: { name: 'foo', pid: '1234' } })
+  t.equal(str, '(foo/1234)')
+})
+
+tap.test('works with only `name` & `hostname` present', async t => {
+  const str = prettifyMetadata({ log: { name: 'foo', hostname: 'bar' } })
+  t.equal(str, '(foo on bar)')
+})
+
+tap.test('works with only `pid` & `hostname` present', async t => {
+  const str = prettifyMetadata({ log: { pid: '1234', hostname: 'bar' } })
+  t.equal(str, '(1234 on bar)')
+})
+
+tap.test('works with only `name`, `pid`, & `hostname` present', async t => {
+  const str = prettifyMetadata({ log: { name: 'foo', pid: '1234', hostname: 'bar' } })
+  t.equal(str, '(foo/1234 on bar)')
+})
+
+tap.test('works with only `name` & `caller` present', async t => {
+  const str = prettifyMetadata({ log: { name: 'foo', caller: 'baz' } })
+  t.equal(str, '(foo) <baz>')
+})
+
+tap.test('works with only `pid` & `caller` present', async t => {
+  const str = prettifyMetadata({ log: { pid: '1234', caller: 'baz' } })
+  t.equal(str, '(1234) <baz>')
+})
+
+tap.test('works with only `hostname` & `caller` present', async t => {
+  const str = prettifyMetadata({ log: { hostname: 'bar', caller: 'baz' } })
+  t.equal(str, '(on bar) <baz>')
+})
+
+tap.test('works with only `name`, `pid`, & `caller` present', async t => {
+  const str = prettifyMetadata({ log: { name: 'foo', pid: '1234', caller: 'baz' } })
+  t.equal(str, '(foo/1234) <baz>')
+})
+
+tap.test('works with only `name`, `hostname`, & `caller` present', async t => {
+  const str = prettifyMetadata({ log: { name: 'foo', hostname: 'bar', caller: 'baz' } })
+  t.equal(str, '(foo on bar) <baz>')
+})
+
+tap.test('works with only `caller` present', async t => {
+  const str = prettifyMetadata({ log: { caller: 'baz' } })
+  t.equal(str, '<baz>')
+})
+
+tap.test('works with only `pid`, `hostname`, & `caller` present', async t => {
+  const str = prettifyMetadata({ log: { pid: '1234', hostname: 'bar', caller: 'baz' } })
+  t.equal(str, '(1234 on bar) <baz>')
+})
+
+tap.test('works with all four present', async t => {
+  const str = prettifyMetadata({ log: { name: 'foo', pid: '1234', hostname: 'bar', caller: 'baz' } })
+  t.equal(str, '(foo/1234 on bar) <baz>')
+})
